feat: Refactor app to use Flask Blueprints
