+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode
 env
 __pycache__
+.pytest_cache
