+++ docker-compose.yml
@@ -12,9 +12,10 @@ services:
     ports:
       - 5001:5000
     environment:
-      - FLASK_APP=project/__init__.py
       - FLASK_ENV=development
       - APP_SETTINGS=project.config.DevelopmentConfig
+      - DATABASE_URL=postgresql://postgres:postgres@fyyur-db:5432/fyyur_dev
+      - DATABASE_TEST_URL=postgresql://postgres:postgres@fyyur-db:5432/fyyur_test
     depends_on:
       - fyyur-db
