+++ manage.py
@@ -1,8 +1,10 @@
 from flask.cli import FlaskGroup
 
-from project import app, db
+from project import create_app, db
+from project.api.models import Artist
 
-cli = FlaskGroup(app)
+app = create_app()
+cli = FlaskGroup(create_app=create_app)
 
 @cli.command('recreate_db')
 def recreate_db():
